feat(Portal): move portal to `core` components
